Some Fixies found by Coverity (mostly leaks) Add ECDHE-RSA-AES256-GCM-SHA384 cipher

+++ dcpp/ADLSearch.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2001-2015 Jacek Sieka, arnetheduck on gmail point com
+ * Copyright (C) 2001-2016 Jacek Sieka, arnetheduck on gmail point com
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -56,8 +56,8 @@ public:
 	
 	// dest dir index
 	unsigned long ddIndex;
-	int64_t adlsRaw;
-	int64_t adlsPoints;
+	int adlsRaw;
+	int adlsPoints;
 	/// Search source type
 	enum SourceType {
 		TypeFirst = 0,
