+++ dcpp/TimerManager.cpp
@@ -100,7 +100,7 @@ uint64_t TimerManager::getTick() {
 	timeval tv2;
 	gettimeofday(&tv2, NULL);
 	/// @todo check conversions to use uint64_t fully
-	return static_cast<uint64_t>(((tv2.tv_sec - tv.tv_sec) * 1000 ) + ( (tv2.tv_usec - tv.tv_usec) / 1000));
+	return static_cast<uint64_t>(((tv2.tv_sec - tv.tv_sec) * static_cast<uint64_t>(1000) ) + ( (tv2.tv_usec - tv.tv_usec) / static_cast<uint64_t>(1000) ));
 #endif
 }
 }//namespace dcpp
